Added DEBUG settings for heater and pwm, renamed temperature to temp to be more consistem with rest of code. Changed from an 8-bits to a 32-bits variable to contain all the extra bits.

+++ debug.c
@@ -1,5 +1,5 @@
 #include	"debug.h"
 #include	"config.h"
 
-volatile uint8_t	debug_flags = DEBUG_INIT;
+volatile uint32_t debug_flags = DEBUG_INIT;
